move code to datasvc, fallback to old style PUT
