+++ cmd/revad/svcs/httpsvcs/datasvc/datasvc.go
@@ -21,10 +21,11 @@ package datasvc
 import (
 	"fmt"
 	"net/http"
-	"os"
 
 	"github.com/cs3org/reva/cmd/revad/httpserver"
 	"github.com/cs3org/reva/cmd/revad/svcs/httpsvcs"
+	tusd "github.com/cs3org/reva/cmd/revad/svcs/httpsvcs/datasvc/handler"
+	"github.com/cs3org/reva/pkg/appctx"
 	"github.com/cs3org/reva/pkg/storage"
 	"github.com/cs3org/reva/pkg/storage/fs/registry"
 	"github.com/mitchellh/mapstructure"
@@ -37,7 +38,6 @@ func init() {
 type config struct {
 	Prefix       string                            `mapstructure:"prefix"`
 	Driver       string                            `mapstructure:"driver"`
-	TmpFolder    string                            `mapstructure:"tmp_folder"`
 	Drivers      map[string]map[string]interface{} `mapstructure:"drivers"`
 	ProviderPath string                            `mapstructure:"provider_path"`
 }
@@ -55,14 +55,6 @@ func New(m map[string]interface{}) (httpsvcs.Service, error) {
 		return nil, err
 	}
 
-	if conf.TmpFolder == "" {
-		conf.TmpFolder = os.TempDir()
-	}
-
-	if err := os.MkdirAll(conf.TmpFolder, 0755); err != nil {
-		return nil, err
-	}
-
 	fs, err := getFS(conf)
 	if err != nil {
 		return nil, err
@@ -72,8 +64,8 @@ func New(m map[string]interface{}) (httpsvcs.Service, error) {
 		storage: fs,
 		conf:    conf,
 	}
-	s.setHandler()
-	return s, nil
+	err = s.setHandler()
+	return s, err
 }
 
 // Close performs cleanup.
@@ -81,6 +73,12 @@ func (s *svc) Close() error {
 	return nil
 }
 
+// Create a new DataStore instance which is responsible for
+// storing the uploaded file on disk in the specified directory.
+// This path _must_ exist before tusd will store uploads in it.
+// If you want to save them on a different medium, for example
+// a remote FTP server, you can implement your own storage backend
+// by implementing the tusd.DataStore interface.
 func getFS(c *config) (storage.FS, error) {
 	if f, ok := registry.NewFuncs[c.Driver]; ok {
 		return f(c.Drivers[c.Driver])
@@ -96,24 +94,82 @@ func (s *svc) Handler() http.Handler {
 	return s.handler
 }
 
-func (s *svc) setHandler() {
-	s.handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		switch r.Method {
-		case "HEAD":
-			addCorsHeader(w)
-			w.WriteHeader(http.StatusOK)
-			return
-		case "GET":
-			s.doGet(w, r)
-			return
-		case "PUT":
-			s.doPut(w, r)
-			return
-		default:
-			w.WriteHeader(http.StatusNotImplemented)
-			return
+func (s *svc) setHandler() (err error) {
+	composable, ok := s.storage.(tusd.Composable)
+	if ok {
+		// A storage backend for tusd may consist of multiple different parts which
+		// handle upload creation, locking, termination and so on. The composer is a
+		// place where all those separated pieces are joined together. In this example
+		// we only use the file store but you may plug in multiple.
+		composer := tusd.NewStoreComposer()
+		// TODO use Terminater
+		// TODO use Locker
+		// TODO use Concater
+		// TODO use LenghtDeferrer
+		composable.UseIn(composer)
+
+		//logger := log.New(os.Stdout, "tusd ", log.Ldate|log.Ltime|log.Lshortfile)
+
+		config := tusd.Config{
+			BasePath:      s.conf.Prefix,
+			StoreComposer: composer,
+			//Logger:        logger,
 		}
-	})
+
+		handler, err := tusd.NewUnroutedHandler(config)
+		if err != nil {
+			return err
+		}
+
+		s.handler = handler.Middleware(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+
+			log := appctx.GetLogger(r.Context())
+			log.Info().Msgf("tusd routing: path=%s", r.URL.Path)
+
+			switch r.Method {
+			// old fashioned download.
+
+			// GET is not part of the tus.io protocol
+			// currently there is no way to GET an upload that is in progress
+			// TODO allow range based get requests? that end before the current offset
+			case "GET":
+				s.doGet(w, r)
+
+			// tus.id based upload
+
+			// uploads are initiated using the CS3 APIs Initiate Download call
+			//case "POST":
+			//	handler.PostFile(w, r)
+			case "HEAD":
+				handler.HeadFile(w, r)
+			case "PATCH":
+				handler.PatchFile(w, r)
+			// TODO Only attach the DELETE handler if the Terminate() method is provided
+			case "DELETE":
+				handler.DelFile(w, r)
+			}
+		}))
+	} else {
+		s.handler = http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+			switch r.Method {
+			case "HEAD":
+				addCorsHeader(w)
+				w.WriteHeader(http.StatusOK)
+				return
+			case "GET":
+				s.doGet(w, r)
+				return
+			case "PUT":
+				s.doPut(w, r)
+				return
+			default:
+				w.WriteHeader(http.StatusNotImplemented)
+				return
+			}
+		})
+	}
+
+	return err
 }
 
 func addCorsHeader(res http.ResponseWriter) {
