+++ cmd/revad/svcs/httpsvcs/datasvc/put.go
@@ -20,11 +20,17 @@ package datasvc
 
 import (
 	"net/http"
+	"strings"
+
+	storageproviderv0alphapb "github.com/cs3org/go-cs3apis/cs3/storageprovider/v0alpha"
+	"github.com/cs3org/reva/pkg/appctx"
 )
 
+// TODO deprecated ... use tus
 func (s *svc) doPut(w http.ResponseWriter, r *http.Request) {
-	/*ctx := r.Context()
+	ctx := r.Context()
 	log := appctx.GetLogger(ctx)
+	log.Warn().Msg("error dumb PUT requests are deprecated, make your storage drive implement tus.io")
 	fn := r.URL.Path
 
 	fsfn := strings.TrimPrefix(fn, s.conf.ProviderPath)
@@ -35,7 +41,7 @@ func (s *svc) doPut(w http.ResponseWriter, r *http.Request) {
 		log.Error().Err(err).Msg("error uploading file")
 		w.WriteHeader(http.StatusInternalServerError)
 		return
-	}*/
+	}
 
 	r.Body.Close()
 	w.WriteHeader(http.StatusOK)
