+++ go.mod
@@ -36,3 +36,5 @@ require (
 	gopkg.in/ldap.v2 v2.5.1
 	gopkg.in/square/go-jose.v2 v2.2.2 // indirect
 )
+
+go 1.13
